Allow customization of imports for GremlinGroovyScriptEngine.
Need to think this concept through in a more global fashion.  This fix only helps gremlin groovy.  It won't help other ScriptEngines.  On the other hand all scriptengines have their own implementation so there may not be any way to enforce this feature across them all anyway.  Would be nice if ScriptEngineManager took some parameters or something that could get shoved into initialized ScriptEngine implementations but that doesn't appear to be an option.  Even with this change, the implementation is really not so nice when using ScriptEngineManager to initialize factories because the import provider has to be configured statically which is not so great.  That's a somewhat common pattern at this point, but there really should be a nicer way.